#64 REFACTOR: 빈 등록 이름을 수정한다
