+++ src/main/java/com/project/zighang/global/client/clova/embedding/EmbeddingGenerator.java
@@ -1,0 +1,7 @@
+package com.project.zighang.global.client.clova.embedding;
+
+import java.util.List;
+
+public interface EmbeddingGenerator {
+    List<Double> embed(String text);
+}
